requirements.txt
telethon==1.34.0
asyncpg==0.29.0
python-dotenv==1.0.0
geopy==2.4.1
pycountry==23.12.11